requirements.txt
pywin32
opencv-python
serial-tool 
pyautogui
mss